Move some modules into gdb/
