+++ docs/source/api/info.rst
@@ -1,5 +1,5 @@
-:mod:`pwndbg.info` --- pwndbg.info
+:mod:`pwndbg.gdblib.info` --- pwndbg.gdblib.info
 =============================================
 
-.. automodule:: pwndbg.info
+.. automodule:: pwndbg.gdblib.info
     :members:
