+++ pwndbg/gdblib/ctypes.py
