+++ pwndbg/gdblib/prompt.py
